arm: dts: imx6q-phytec-pbab01: adjust weim pin-muxing and add reg-shift for smsc
Reason for reg-shift = <1>:

The connected address lines between the I.MX6 and the SMSC devices
start at EIM_AD01, not EIM_AD00 so the intention is to tell
the SMSC to shift the reads and writes by 1 to align its output
with the CPU's input.

SMSC (base 1) <--> I.MX6 (base 0) hardware mapping
         A7   <--> AD07
         A6   <--> AD06
         A5   <--> AD05
         A4   <--> AD04
         A3   <--> AD03
         A2   <--> AD02
         A1   <--> AD01
         XX   <--> AD00

Additionally, since FIFO_SEL is pulled high with pin-mux,
SMSC A[7:3] lines are ignored.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>